requirements.txt
Django==4.2
mysqlclient==2.2.0
django-environ==0.11.2
celery==5.3.6
redis==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
xhtml2pdf==0.2.11